+++ app/src/main/java/ksc/campus/tech/kakao/map/data/repositoryimpls/SearchResultRepositoryImpl.kt
@@ -3,7 +3,7 @@ package ksc.campus.tech.kakao.map.data.repositoryimpls
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import ksc.campus.tech.kakao.map.data.datasources.KakaoSearchService
-import ksc.campus.tech.kakao.map.data.mapper.SearchResponseMapper
+import ksc.campus.tech.kakao.map.data.mapper.KakaoSearchDtoMapper
 import ksc.campus.tech.kakao.map.domain.models.SearchResult
 import ksc.campus.tech.kakao.map.domain.repositories.SearchResultRepository
 import javax.inject.Inject
@@ -16,7 +16,7 @@ class SearchResultRepositoryImpl @Inject constructor(
     override val searchResult: Flow<List<SearchResult>>
         get() = kakaoSearchService.searchResult(text, key, BATCH_COUNT).map { document ->
             document.map {
-                SearchResponseMapper.mapSearchResponseToSearchResult(it)
+                KakaoSearchDtoMapper.mapSearchResponseToSearchResult(it)
             }
         }
